chat.py
from flask import Flask, Blueprint, request, redirect, jsonify, make_response, render_template
from werkzeug.security import generate_password_hash, check_password_hash
from db import db_connect
from auth import require_auth_token
from flask_socketio import SocketIO, join_room
import datetime
import jwt
import psycopg2
import sys
